--- Capstone - Team 1/Assets/AI/AIBot
@@ -1,17 +1,0 @@
-#START:
-#Pickle to save
-#tensorflow from tensorflow.keras import layers
-
-#Deep Reinforcement Learning or Value Iteration (VI)
-
-
-#Evaluate a given 5x5 grid
-    #Know what piece you are (X or O)
-    #Reward more for controlling middle +(25, 15, 5)
-    #Control Corners +(20)
-
-#GetBestMove
-    #Should take a board, team to move for, and difficulty
-
-#https://neptune.ai/blog/saving-trained-model-in-python#:~:text=To%20save%20a%20deep%20learning,loaded%20later%20to%20make%20predictions.
-#https://www.activestate.com/resources/quick-reads/how-to-create-a-neural-network-in-python-with-and-without-keras/
